server/requirements.txt
pyzmq
python-dotenv
pygame
flask